+++ ICD.Connect.Settings.Tests/AbstractTestOriginator.cs
@@ -15,9 +15,8 @@ namespace ICD.Connect.Settings.Tests
 	public abstract class AbstractTestOriginator<TSettings> : IOriginator<TSettings>
 		where TSettings : ISettings
 	{
-		public event EventHandler OnSettingsClearing;
-		public event EventHandler OnSettingsCleared;
-		public event EventHandler OnSettingsApplied;
+		private eLifecycleState m_LifecycleState;
+		public event EventHandler<LifecycleStateEventArgs> OnLifecycleStateChanged;
 		public event EventHandler OnNameChanged;
 		public event EventHandler<BoolEventArgs> OnDisableStateChanged;
 
@@ -32,6 +31,8 @@ namespace ICD.Connect.Settings.Tests
 		/// Gets the category for this originator type (e.g. Device, Port, etc)
 		/// </summary>
 		public abstract string Category { get; }
+
+		public eLifecycleState LifecycleState => m_LifecycleState;
 
 		public int Id { get; set; }
 
@@ -98,6 +99,14 @@ namespace ICD.Connect.Settings.Tests
 		}
 
 		public void ClearSettings()
+		{
+			throw new NotImplementedException();
+		}
+
+		/// <summary>
+		/// Run Settings - called after all settings are applied
+		/// </summary>
+		public void StartSettings()
 		{
 			throw new NotImplementedException();
 		}