+++ ICD.Connect.Settings/Cores/AbstractCore.cs
@@ -1,4 +1,5 @@
 using System;
+using ICD.Common.Properties;
 using ICD.Connect.Settings.Originators;
 
 namespace ICD.Connect.Settings.Cores
@@ -114,9 +115,10 @@ namespace ICD.Connect.Settings.Cores
 		/// <summary>
 		/// Loads settings from disk and updates the Settings property.
 		/// </summary>
-		public void LoadSettings()
+		/// <param name="postApplyAction">Action that gets called after settings are applicd, before they are started</param>
+		public void LoadSettings(Action postApplyAction)
 		{
-			FileOperations.LoadCoreSettings<AbstractCore<TSettings>, TSettings>(this);
+			FileOperations.LoadCoreSettings<AbstractCore<TSettings>, TSettings>(this, postApplyAction);
 		}
 
 		#endregion