+++ ICD.Connect.Settings/Originators/AbstractOriginator.cs
@@ -19,19 +19,9 @@ namespace ICD.Connect.Settings.Originators
 		where T : ISettings, new()
 	{
 		/// <summary>
-		/// Called when the settings start clearing.
+		/// Raised when the lifecycle state of the originator changes
 		/// </summary>
-		public event EventHandler OnSettingsClearing;
-
-		/// <summary>
-		/// Raised when settings have been cleared.
-		/// </summary>
-		public event EventHandler OnSettingsCleared;
-
-		/// <summary>
-		/// Raised when settings have been applied to the originator.
-		/// </summary>
-		public event EventHandler OnSettingsApplied;
+		public event EventHandler<LifecycleStateEventArgs> OnLifecycleStateChanged;
 
 		/// <summary>
 		/// Raised when the name changes.
@@ -51,6 +41,7 @@ namespace ICD.Connect.Settings.Originators
 		private ICore m_CachedCore;
 		private string m_Name;
 		private bool m_Disable;
+		private eLifecycleState m_LifecycleState;
 
 		#region Properties
 
@@ -63,6 +54,24 @@ namespace ICD.Connect.Settings.Originators
 		/// Gets the category for this originator type (e.g. Device, Port, etc)
 		/// </summary>
 		public abstract string Category { get; }
+
+		/// <summary>
+		/// The lifecycle state of the originator
+		/// Describes the state in the instantiate/load/start/clear/dispose lifecycle
+		/// </summary>
+		public eLifecycleState LifecycleState
+		{
+			get { return m_LifecycleState; }
+			private set
+			{
+				if (m_LifecycleState == value)
+					return;
+
+				m_LifecycleState = value;
+
+				OnLifecycleStateChanged.Raise(this, new LifecycleStateEventArgs(m_LifecycleState));
+			}
+		}
 
 		/// <summary>
 		/// Unique ID for the originator.
@@ -256,9 +265,8 @@ namespace ICD.Connect.Settings.Originators
 		/// <param name="disposing"></param>
 		private void Dispose(bool disposing)
 		{
-			OnSettingsClearing = null;
-			OnSettingsCleared = null;
-			OnSettingsApplied = null;
+			LifecycleState = eLifecycleState.Disposed;
+			OnLifecycleStateChanged = null;
 			OnNameChanged = null;
 			OnDisableStateChanged = null;
 
@@ -342,6 +350,8 @@ namespace ICD.Connect.Settings.Originators
 		{
 			ClearSettings();
 
+			LifecycleState = eLifecycleState.Loading;
+
 			Id = settings.Id;
 			Uuid = settings.Uuid == default(Guid) ? OriginatorUtils.GenerateUuid(Core, Id) : settings.Uuid;
 			Name = string.IsNullOrEmpty(settings.Name) ? GetType().Name : settings.Name;
@@ -355,7 +365,7 @@ namespace ICD.Connect.Settings.Originators
 
 			ApplySettingsFinal(settings, factory);
 
-			OnSettingsApplied.Raise(this);
+			LifecycleState = eLifecycleState.Loaded;
 		}
 
 		/// <summary>
@@ -372,7 +382,7 @@ namespace ICD.Connect.Settings.Originators
 		/// </summary>
 		public void ClearSettings()
 		{
-			OnSettingsClearing.Raise(this);
+			LifecycleState = eLifecycleState.Clearing;
 
 			ClearSettingsFinal();
 
@@ -389,7 +399,7 @@ namespace ICD.Connect.Settings.Originators
 
 			SetPermissions(Enumerable.Empty<Permission>());
 
-			OnSettingsCleared.Raise(this);
+			LifecycleState = eLifecycleState.Cleared;
 		}
 
 		/// <summary>
@@ -443,6 +453,32 @@ namespace ICD.Connect.Settings.Originators
 			}
 
 			ApplySettings((T)settings, factory);
+		}
+
+		/// <summary>
+		/// Start settings for the origintor
+		/// This should be used to start communications with devices and perform initial actions
+		/// </summary>
+		public void StartSettings()
+		{
+			//Only Start Once
+			if (LifecycleState == eLifecycleState.Started)
+				return;
+
+			LifecycleState = eLifecycleState.Starting;
+
+			StartSettingsFinal();
+			
+			LifecycleState = eLifecycleState.Started;
+		}
+
+		/// <summary>
+		/// Override to add actions on StartSettings
+		/// This should be used to start communications with devices and perform initial actions
+		/// </summary>
+		protected virtual void StartSettingsFinal()
+		{
+			
 		}
 
 		#endregion