+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
  - Removed ISimplOriginator
 
+### Removed
+ - Removed redundant telemetry property for system online state - it's now handled by the telemetry provider
+
 ## [11.1.0] - 2020-07-14
 ### Changed
  - Reverted IdUtils for consistency with DAV
@@ -34,6 +37,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Fixed a bug where group settings would serialize references to non-serialized children
  - Originator names default to the type name
  - Config header CompiledOn element is serialized/deserialized as a DateTime
+
+## [10.1.0] - 2020-10-06
+### Changed
+ - Added LifecycleState property/event to IOriginator, and removed individual events
 
 ## [10.0.0] - 2020-03-20
 ### Added